requirements.txt
customtkinter==5.2.2
Pillow==11.0.0
psutil==6.1.1
pystray==0.19.5
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
wgconfig==1.0.4
